rename logs1 to logs
